requirements.txt
pyglet
numpy
matplotlib
Pillow
